algo finished, there may be bugs but nothing huge, going to start a new algorithm

+++ srcs/Node.class.cpp
@@ -112,7 +112,7 @@ void        Node::closeNode(void)
         
         bool    checkIfOpened = false;
         int     bestidx = -1;
-        int     bestval = INT_MAX;
+        double  bestval = INT_MAX;
         for (int j = 0; j < this->size * this->size; ++j)
         {
             if (prev->tab[j] != nullptr)
